vrc6brat: Numerous changes.
* Move ports around to support physical build.
* Add digital sample output with sample sample, currently from capslock.mod and plays repeatedly.

+++ vrc6brat/ports.h
@@ -1,6 +1,6 @@
 #ifndef _PORTS_H_
 #define _PORTS_H_
 
-void port_write(uint16_t addr, uint8_t data);
+void port_write(uint8_t addr, uint8_t data);
 
 #endif
